Nicolas Regnault: - add support for hamiltonian precalculation in TensorProductSparseMatrixSelectedBlockHamiltonian - add class for generaic hamiltonian precalculation operation - start cleaning some deprecated code
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@2437 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/Hamiltonian/ApproximateFermionOnTorusCoulombHamiltonian.cc
@@ -672,7 +672,7 @@ double ApproximateFermionOnTorusCoulombHamiltonian::PartialMisraFunction (double
 //
 // return value = amount of memory needed
 
-int ApproximateFermionOnTorusCoulombHamiltonian::FastMultiplicationMemory()
+long ApproximateFermionOnTorusCoulombHamiltonian::FastMultiplicationMemory()
 {
   int Index;
   int ReducedMaxMomentum = this->MaxMomentum - 1;
