+++ FQHE/src/Hamiltonian/ApproximateFermionOnTorusCoulombHamiltonian.h
@@ -267,7 +267,7 @@ class ApproximateFermionOnTorusCoulombHamiltonian : public AbstractHamiltonian
   // test the amount of memory needed for fast multiplication algorithm
   //
   // return value = amount of memory needed
-  int FastMultiplicationMemory();
+  long FastMultiplicationMemory();
 
   // enable fast multiplication algorithm
   //
