+++ FQHE/src/Hamiltonian/ParticleOnCircleCoulombHamiltonian.h
@@ -251,7 +251,7 @@ class ParticleOnCircleCoulombHamiltonian : public AbstractHamiltonian
   // test the amount of memory needed for fast multiplication algorithm
   //
   // return value = amount of memory needed
-  int FastMultiplicationMemory();
+  long FastMultiplicationMemory();
 
   // enable fast multiplication algorithm
   //
