+++ FQHE/src/Hamiltonian/ParticleOnTorusCoulombDMRGReadyHamiltonian.h
@@ -257,7 +257,7 @@ class ParticleOnTorusCoulombDMRGReadyHamiltonian : public AbstractHamiltonian
   // test the amount of memory needed for fast multiplication algorithm
   //
   // return value = amount of memory needed
-  int FastMultiplicationMemory();
+  long FastMultiplicationMemory();
 
   // enable fast multiplication algorithm
   //
